bot_service/pyproject.toml
[project]
name = "bot-service"
version = "0.1.0"
description = "Telegram bot for LLM consultations"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.112.0",
    "uvicorn[standard]>=0.30.0",
    "aiogram>=3.0.0",
    "sqlalchemy>=2.0.30",
    "aiosqlite>=0.20.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "aiohttp-socks>=0.11.0",
    "aio-pika>=9.3.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]


[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.14.14",
]